mpthreadport.c: Refactor threading to a working state
- call vTaskSuspend() instead of vTaskDelete()
  vTaskDelete() assumes, that the stack buffer was allocated within
  tasks.c. But the stack buffer is allocated form the MP heap. Freeing
  it the wrong way causes memory corruption.
- have all threads the same priority. Otherwise the following threads
  will execute with much slower pace.
- Remove usage of thread_cnt. It was only used to give threads
  different priorities, which seemed to be a bad decision anyhow.
- Do a hard reset with Ctrl-D if threads had been started.
  That will close all tasks and free the heap.
- call tls_os_time_delay(1) in the MICROPY_EVENT_POLL_HOOK macro
  giving time for a task switch.

Note:
This variant provides the stack for the threads (= RTOS tasks) from the
micropython heap. That prohibits using xTaskDelete(), because that
assumes, that the stack is provided by the RTOS heap, an xTaskDelete()
frees that using RTOS methods, causing the both heaps to get corrupted.
When using xTaskSuspend is is unclear, whether the thread stack buffer
can be released. At the moment, it is kept.
I can provide another variant, which is taking the task stack buffer
from the RTOS heap. Since that is very limited, only one thread can be
started at 4k stack size, or two threads at 2k stack size. The RTOS
heap can be increased by changing the setting in FreeRTOSConfig.h. But
that is a change to the SDK I want to avoid.