esp8266 docs: Add quickref documentation for UART on esp8266
This patch adds quickref documentation for the change in commit
afd0701bf7a9dcb50c5ab46b0ae88b303fec6ed3.  This commit added the ability to
disable the REPL and hence use UART0 for serial communication on the esp8266,
but was not previously documented anywhere.

The text is largely taken from the commit message, with generic information on
using the UART duplicated from the Wipy quickref document.