actions: Try checkout v2.2.0
This version is supposed to
> Fetch all history for all tags and branches when fetch-depth=0

We leave the tags fetch in place so that actions _in cloned repos_
work.  Cloned repos' tags do not necessarily match adafruit/circuitpython,
but we want version reporting to be able to use adafruit/circuitpython
tags when they are most relevant according to "git describe"'s heuristics.

Submodules are different, as they always point to the repo specified
in .gitmodules, so they don't need special handling to get the most
relevant tags.