py: Rename sys module to usys.
This is consistent with the other 'micro' modules and allows
implementing additional features via e.g. micropython-lib's sys.
Note this change is not backwards compatible for ports which do
not enable weak links as they used to be able to import sys but
won't be able to do that anymore now, and need to use import usys.