nrf/bluetooth: Handle data length update request.
Bluetooth link gets disconnected when connecting from a PC
after 30-40 seconds.

This patch add handling of the data length update request.
The data length parameter pointer is set to NULL in the reply.
This will let the SoftDevice automatically set values and
use them in the data length update proceedure.