travis: Add zephyr build to CI.
Adds a test to build the zephyr port in CI using the same docker
container that the zephyr project uses for its own CI.

Always make clean zephyr builds to ensure we don't just rebuild C code,
but we also rebuild Kconfig and dts. This is required when switching
between boards, which have different Kconfigs and device trees.

Uses the tagged zephyr 2.3.0 release.

Signed-off-by: Maureen Helm <maureen.helm@nxp.com>