mpy-cross: Silently clone when building for static-raspbian
Also move the construct earlier.  This prevents the message
```
make: pitools/arm-bcm2708/gcc-linaro-arm-linux-gnueabihf-raspbian-x64/bin/arm-linux-gnueabihf-gcc: Command not found
```
from occurring during the first build, since make was evaluating the line
```
mpy-cross.mk:COMPILER_TARGET := $(shell $(CC) -dumpmachine)
```
before $(CC) was created by the clone.  Nothing bad seems to have come of
this, but it's nice to fix it anyhow.  (And interesting that it was lost
among the spewed messages we're removing now)