extmod/nimble: Add timeout for HCI sync on startup.
This allows `ble.active(1)` to fail correctly if the HCI controller is
unavailable.

Also avoids an infine loop in the NimBLE event handler where NimBLE doesn't
correctly detect that the HCI controller is unavailable and keeps trying to
reset.

Also fixes an issue where GATT service registrations were left allocated,
which led to a bad realloc if the stack was activated multiple times.