extmod/machine_i2c: Fix buffer overrun, if 'addrsize' is bigger than 32
The memory operation functions read_mem() and write_mem() create a temporary
buffer on the local stack for the address bytes with the size of 4 bytes.
This buffer is filled in a loop from the user supplied address and address length.
If the user supplied 'addrsize' is bigger than 32, the local buffer is overrun.

Fix this by raising an exception for invalid 'addrsize' values.

Signed-off-by: Michael Buesch <m@bues.ch>