zephyr: Use cmake ExternalProject to build libmicropython.a.
Rework the zephyr port's build infrastructure to build libmicropython.a
with a cmake ExternalProject, instead of invoking cmake from a Makefile
(which was an awkward thing to do because cmake typically generates
Makefiles).

This change makes it possible to build the zephyr port like most other
zephyr applications using west or cmake directly. It simplifies building
with extra cmake arguments, such as specifying an alternate conf file or
adding an Arduino shield. It also enables building the zephyr port
anywhere in the host file system, which will allow regressing across
multiple boards with the zephyr sanitycheck script.

To build with west:

    $ west build -b frdm_k64f ~/micropython/ports/zephyr

To build with cmake:

    $ mkdir build && cd build
    $ cmake -DBOARD=frdm_k64f ~/micropython/ports/zephyr
    $ make

To build the minimal configuration:

    $ west build -b frdm_k64f -- -DCONF_FILE=prj_minimal.conf

To build with a zephyr-supported Arduino shield:

    $ west build -b frdm_k64f -- -DSHIELD=frdm_kw41z

Signed-off-by: Maureen Helm <maureen.helm@nxp.com>