windows/msvc: Support freezing modules.
Support freezing modules via manifest.py for consistency with the
other ports.  In essence this comes down to calling makemanifest.py
and adding the resulting .c file to the build.  Note the file with
preprocessed qstrs has been renamed to match what makemanifest.py
expects and which is also the name all other ports use.