py/gc: Allow allocations to survive soft-resets.
This allows gc_set_permanant(ptr, true) to allow a given pointer to be treated
as "always marked" by the gc.

The plan for this feature is to allow NimBLE (and btstack) to allocate on the
heap in a way that survives a soft reset (allowing us to decouple stack
initialisation from MicroPython initialisation). The next step will be to make
the ports no longer gc_init() after each soft reset, instead relying only on
gc_sweep_all.