py/makeqstrdefs.py: Support preprocessing C++ files for QSTR generation.
When SCR_QSTR contains C++ files they should be preprocessed
with the same compiler flags (CXXFLAGS) as they will be
compiled with, to make sure code scanned for QSTR occurrences
is effectively the code used in the rest of the build.
The 'split SCR_QSTR in .c and .cpp files and process each with
different flags' logic isn't trivial to express in a Makefile
and the existing principle for deciding which files to preprocess
was already rather complicated, so the actual preprocessing is
moved into makeqstrdefs.py completely.