py/py.mk: Support C++ code for user C modules.
Support C++ code in .cpp files by providing CXX counterparts of
the _USERMOD_ flags we have for C already.  This merely enables
the Makefile of user C modules to use variables specific to C++
compilation, it is still up to each port's main Makefile to also
include these in the build.