esp32, unix: Support building C++ code.
Support building .cpp files and linking them into the micropython
executable in a way similar to how it is done for .c files.  The main
incentive here is to enable user C modules to use C++ files (which are
put in SRC_MOD_CXX by py.mk) since the core itself does not utilize C++.
However to verify build functionality a unix overage test is added.
The esp32 port already has CXXFLAGS so just add the user modules' flags
to it, for the unix port use a copy of the CFLAGS but strip the ones
which are not usable for C++.