examples: Add user C modules.
Add working example code to provide a starting point for users with
files the can just copy, and include the modules in the coverage test
to verify the complete user C module build functionality.  The cexample
module uses the code originally found in cmodules.rst, which has been
updated to reflect this and partially rewritten with more complete
information.