ports/stm32/Makefile: Make the generation of `firmware.bin` explicit
The file `$(BUILD)/firmware.bin` was used by the target `deploy-stlink`
and `deploy-openocd` but it was generated indirectly by the target
`firmware.dfu`.

As this file could be used to program boards directly by a Mass Storage
copy, it's better to make it explicitly generated.

Some target are refactored to remove redundancy and be more explicit on
dependencies.