special_accessors: runtime addition of descriptors could set
Adding descriptors at runtime doesn't set the special accessor flag.
or atleast the way I was doing it.

type_attr calls `check_for_special_accessors` (before adding the special accessor flag)  which will obviously return false (because no special accessor was added
on the original object)

I've replaced the validation to check if the new attribute to be added is an descriptor (if true - then set the special accessor flag)

Signed-off-by: Jonathan Bruchim <yonbruchim@gmail.com>