tests/extmod: also try to mount a block device directly.
Mounting a bdev directly tries to auto-detect the filesystem and if none is found an OSError(19,) should be raised.
The fourth parameter of readblocks() and writeblocks() must be optional to support ports with MICROPY_VFS_FAT=1. Otherwise mounting a bdev may fail because looking for a FATFS will call readblocks() with only 3 parameters.