tests/misc/sys_settrace: Make test output independent of invoked path.
The original logic of reducing a full path to a relative one assumes
"tests/misc" is in the filename which is limited in usage: it never
works for CPython on windows since that will use a backslash as path
separator, and also won't work when the filename is a path not relative
to the tests directory which happens for example in the common case of
running "./run-tests -d misc".
Fix all cases by printing only the bare filename, which requires them
all to start with sys_settrace_ hence the renaming.