extmode/modurandom.c: getrandombits() method has proper error message. tests/cpydiff: Random module differences for getrandbits and randint. tests/cpydiff: Int type describes lack of bit length.
modurandom.c
The random module's getrandbits() method didn't give a proper error
message when calling it with a value that was outside of the range
of 1-32. Now instead of simply giving a ValueError it gives an error
message that states that the number of bits must be between 1 and 32.

Random Module Tests
Since the random module's methods getrandbits() and randint() differe
from the CPython random module's version of the methods tests have been
added. For getrandbits the relevant documentation is shown and added to
the docs. The same is given for randint method so that the information
is more easily found. Further there is an example function showing a
possible workaround to get more than 32bits of data in a simple way.

Int Object's lack of bit_length() method.
Finally since the Int object lacks the bit_length() method ther is a test
for that method also to include within the docs. It now shows the person
what they can see on micropython compared to CPython.